Code formatting, moved some functions to the utils.indra module
